+++ R/zzz.R
@@ -1,48 +1,9 @@
 ## Path for different operating system
 OS <- Sys.info()["sysname"]
-sysDrive <- switch(OS,
-  Linux = "/mnt/F",
-  Windows = "F:"
-)
 
 ## Options for Folkehelseprofiler
-opt.orgdata <- list(
-  orgdata.year = NULL, # which year of geo codes for recoding should refer to
-  orgdata.drive = sysDrive,
-  orgdata.folder.db = "Forskningsprosjekter/PDB 2455 - Helseprofiler og til_/PRODUKSJON/STYRING/raw-khelse",
-  orgdata.folder.data = "F:/Forskningsprosjekter/PDB 2455 - Helseprofiler og til_/PRODUKSJON/ORGDATA",
-  orgdata.db = "raw-database_BE.accdb",
-  orgdata.geo = "geo-koder.accdb",
-  orgdata.verbose = TRUE,
-  orgdata.aggregate = TRUE,
-  orgdata.implicit.null = TRUE,
-  orgdata.recode.base = FALSE,
-  orgdata.parallel = FALSE,
-
-  ## Number of TABS and VALS
-  orgdata.tabs = 3,
-  orgdata.vals = 3,
-
-  ## Standard columns
-  orgdata.columns = c("GEO", "LEVEL", "AAR", "KJONN", "ALDER", "UTDANN", "LANDSSB"),
-
-  ## Columns with numeric values. GEO must not be converted to numeric at early
-  ## stage because some geo codes has to be a combination of 2 columns and need
-  ## to keep the leading 0 on second column before merging. GEO will only be
-  ## converted to numeric after merging
-  orgdata.num = c("GEO", "AAR", "KJONN", "ALDER"),
-
-  ## Use for debuging purposes only
-  orgdata.debug = FALSE,  #Choose "deep" to show helper functions as well
-  orgdata.debug.nrow = 0,  #If other than 0 then x rows will be selected
-  orgdata.debug.row = NULL, #Select specific rows eg. 20:50
-  orgdata.debug.aggregate = FALSE, #Recode but don't aggregate if TRUE
-  orgdata.debug.geo = FALSE, #Keep old geo if TRUE
-
-  ## Just for fun
-  orgdata.emoji = "mark" #can choose mark, thumb, smile or sad
-)
-
+optOrg <- yaml::yaml.load_file("https://raw.githubusercontent.com/helseprofil/config/main/config-orgdata.yml")
+opt.orgdata <- as.list(opt_rename(optOrg))
 
 .onLoad <- function(libname, pkgname) {
   op <- options()
@@ -52,5 +13,5 @@ opt.orgdata <- list(
 }
 
 .onAttach <- function(libname, pkgname) {
-  packageStartupMessage("orgdata version 0.5.7")
+  packageStartupMessage("orgdata version 0.5.8")
 }