+++ tests/testthat/test-read-file.R
@@ -1,4 +1,9 @@
 test_that("Read rawdata", {
+
+  dtOut <- readRDS(system.file("testdata", "testfile_read.rds", package = "orgdata"))
+
+  expect_equal(read_file(file.path(system.file(package = "orgdata"), "testdata/testfile.xlsx"), skip = 1, range = "A2:B6"), dtOut)
+
   expect_error(find_data("raw-data.spss"))
   expect_error(read_file.default("raw-data.spss"))
   expect_error(read_file("raw-data.spss"))
@@ -26,7 +31,19 @@ test_that("File path", {
   fcsv <- "C:\\TEST\\file.csv"
   fcsv2 <- "C:/TEST/file.csv"
 
-  expect_equal(is_file_path(fcsv), fcsv)
-  expect_equal(is_file_path(fcsv2), fcsv2)
+  expect_equal(is_read_path(fcsv), fcsv)
+  expect_equal(is_read_path(fcsv2), fcsv2)
   expect_error(read_file(file = system.file("testdata", "bla.csv", package = "orgdata")))
 })
+
+
+test_that("Dots args", {
+
+  Dots <- list(trimws = TRUE, nrows = 50)
+  csvOut <- list(strip.white = TRUE, nrows = 50)
+  xlsOut <- list(trim_ws = TRUE, n_max = 50)
+
+  expect_equal(is_csv_dots(Dots), csvOut)
+  expect_equal(is_xls_dots(Dots), xlsOut)
+  })
+